// types/global.d.ts
export type MsgObj = WAMessage & {
  
  id:              string
  isBaileys:       boolean
  chat:            string
  fromMe:          boolean
  isGroup:         boolean
  sender:          string
  mtype:           string
  msg:             proto.IMessage[keyof proto.IMessage] | null | undefined
  body:            string
  text:            string
  mentionedJid:    string[]
  quoted:          MsgObj | null
  
  reply:           (text: string | Buffer, chatId?: string, opts?: Record<string, unknown>) => Promise<WAMessage | undefined>
  download:        () => Promise<Buffer>
  copy:            () => MsgObj
  copyNForward:    (jid?: string, force?: boolean, opts?: Record<string, unknown>) => Promise<WAMessage | undefined>
  getQuotedObj:    () => Promise<MsgObj | false>
  getQuotedMessage: () => Promise<MsgObj | false>
  
  _isJadibot?:      boolean
  _stikerHandled?:  boolean
}


export type Message = MsgObj


export type GetFileResult = {
  res?:      Buffer | undefined
  filename:  string
  size:      string | number
  ext:       string
  mime:      string
  data:      Buffer
  cleanup:   () => void
}

export type ExtSocket = WASocket & {
  
  public?:                  boolean
  
  decodeJid:                (jid: string) => string
  getFile:                  (path: string | Buffer, save?: boolean) => Promise<GetFileResult>
  downloadMediaMessage:     (message: MsgObj | Record<string, unknown>) => Promise<Buffer>
  downloadMedia:            (message: MsgObj | Record<string, unknown>, opts?: {
    forceType?: string | null
    maxRetries?: number
    retryDelay?: number
  }) => Promise<Buffer>
  sendText:                 (jid: string, text: string, quoted?: unknown, opts?: Record<string, unknown>) => Promise<WAMessage | undefined>
  sendImageAsSticker:       (jid: string, src: string | Buffer, quoted?: unknown, opts?: Record<string, unknown>) => Promise<Buffer>
  sendVideoAsSticker:       (jid: string, src: string | Buffer, quoted?: unknown, opts?: Record<string, unknown>) => Promise<Buffer>
  downloadAndSaveMediaMessage: (message: MsgObj | Record<string, unknown>, filename: string, attachExtension?: boolean) => Promise<string>
  sendMedia:                (jid: string, path: string, caption?: string, quoted?: unknown, opts?: Record<string, unknown>) => Promise<WAMessage | undefined>
  sendPoll:                 (jid: string, question: string, options: string[]) => Promise<WAMessage | undefined>
}



export type FkontakMsg = {
  key: {
    participant: string
    fromMe:      boolean
    id:          string
    remoteJid:   string
  }
  message: {
    contactMessage: {
      displayName:    string
      vcard:          string
      jpegThumbnail?: Buffer
    }
  }
}


export type HandleData = {
  Morela:                      ExtSocket
  conn:                        ExtSocket
  text:                        string
  args:                        string[]
  isOwn:                       boolean
  isPrem:                      boolean
  isAdmin:                     boolean
  botAdmin:                    boolean
  senderJid:                   string
  usedPrefix:                  string
  command?:                    string
  reply:                       (msg: string, opt?: Record<string, unknown>) => Promise<unknown>
  fkontak:                     FkontakMsg | MsgObj   
  downloadContentFromMessage?: unknown
}


export type PluginHandler = {
  (m: Message, ctx: HandleData): Promise<unknown>
  command?:   string[]
  tags?:      string[]
  help?:      string[]
  owner?:     boolean
  premium?:   boolean
  group?:     boolean
  private?:   boolean
  admin?:     boolean
  botAdmin?:  boolean
  noLimit?:   boolean
  passive?:   boolean
}

export type PluginModule = {
  handler:    PluginHandler | ((m: Message, ctx: HandleData) => Promise<unknown>)
  command?:   string[]
  tags?:      string[]
  help?:      string[]
  owner?:     boolean
  premium?:   boolean
  group?:     boolean
  private?:   boolean
  admin?:     boolean
  botAdmin?:  boolean
  noLimit?:   boolean
  passive?:   boolean
}